+++ src/containers/WorkspaceMounted/Collection/index.jsx
@@ -137,31 +137,25 @@ class Collection extends React.Component{
   }
 
   setCreateItemView(){
-    //service.api.parentTempHideMobilePreview();
     this.setState({view:{key:'createItem', item: null}, modalBusy:false});
   }
 
   setRenameItemView(item: any){
-    //service.api.parentTempHideMobilePreview();
     this.setState({view:{key:'renameItem', item}, modalBusy:false});
   }
   setCopyItemView(item: any){
-    //service.api.parentTempHideMobilePreview();
     this.setState({view:{key:'copyItem', item}, modalBusy:false});
   }
 
   setMakePageBundleItemView(item: any){
-    //service.api.parentTempHideMobilePreview();
     this.setState({view:{key:'makePageBundleItem', item}, modalBusy:false});
   }
 
   setDeleteItemView(item: any){
-    //service.api.parentTempHideMobilePreview();
     this.setState({view:{key:'deleteItem', item }, modalBusy:false});
   }
 
   setRootView(){
-    //service.api.parentTempUnHideMobilePreview();
     this.setState({view:undefined, modalBusy:false});
   }
 
@@ -213,7 +207,6 @@ class Collection extends React.Component{
         this.setState({modalBusy:false, view: undefined});
       });
 
-    //service.api.parentTempUnHideMobilePreview();
   }
 
 
@@ -237,13 +230,11 @@ class Collection extends React.Component{
         this.setState({modalBusy:false, view: undefined});
       });
 
-    //service.api.parentTempUnHideMobilePreview();
   }
 
   renameCollectionItem(itemKey : string, itemOldKey: string){
     let { siteKey, workspaceKey, collectionKey } = this.props;
-    if(this.state.view==null)return;
-    //let view = this.state.view;
+    if(this.state.view==null) return;
     service.api.renameCollectionItem(siteKey, workspaceKey, collectionKey, itemOldKey, itemKey)
       .then((result)=>{
         if(result.renamed){
@@ -261,7 +252,6 @@ class Collection extends React.Component{
         this.setState({modalBusy:false, view: undefined});
       });
 
-    //service.api.parentTempUnHideMobilePreview();
   }
 
   copyCollectionItem(itemKey : string, itemOldKey: string){
@@ -285,7 +275,6 @@ class Collection extends React.Component{
         this.setState({modalBusy:false, view: undefined});
       });
 
-    //service.api.parentTempUnHideMobilePreview();
   }
 
   createCollectionItemKey(itemKey : string, itemTitle : string){
@@ -307,7 +296,6 @@ class Collection extends React.Component{
         this.history.push(path);
       });
 
-    //service.api.parentTempUnHideMobilePreview();
   }
 
   resolveFilteredItems = (items: Array<any>) => {
@@ -376,7 +364,6 @@ class Collection extends React.Component{
 
   render(){
 
-    //let { siteKey, workspaceKey, collectionKey } = this.props;
     let { collectionKey } = this.props;
     let { filteredItems, trunked } = this.state;
     let dialog = undefined;
@@ -444,11 +431,6 @@ class Collection extends React.Component{
       />
       }
     }
-
-
-
-    //let pageUrl = this.generatePageUrl(collection);
-    //service.api.updateMobilePreviewUrl(pageUrl)
 
     return(<Route render={ ({history}) => {
       this.history = history;